Separated all the hl states to their own modules
